requirements.txt
numpy
requests
openai==1.3.6
python-dotenv
psutil==5.9.6
pytz
waitress
config
pytesseract==0.3.10
pillow==10.2.0
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80






